We're finally in beta, added two commands "/dona" and "/aiuto"
After few days I'm finally proud to announce that the bot is finished. I maybe have to change the message configuration but finally the bot sends the "correct" message to each user. I have to thank Filippo (who basically created the database) and also Niroge (who helped me to create the code for sending the message). I've added two commands, "/aiuto" which shows you the full list of the commands for the bot and "/dona" that gives you the possibility to donate me some money via PayPal, in order to support my work and to help me to repay the raspberry that hosts the bot, (it's not mandatory, but if you want to be kind...) oh, and don't forget... MERRY CHRISTMAS AND HAPPY NEW YEAR (on the 1 st Jan 2019 I'll finish the beta testing and I'll release the link for the bot) ~ WAPEETY